// app/types/station.ts
export interface Station {
    id: string;
    name: string;
    description: string;
    description2: string;
    coordslatitude: string;
    coordslongitude: string;
    address: string;
    status: string;
    accessTime: string;
    info: string;
    plugCount: string;
    ac: string;
    dcgreaterthan50: string;
    brandid: string;
    lat?: number;
    lng?: number;
} 